+++ Stradario/Strutture/Nodo.cs
@@ -7,22 +7,22 @@ namespace Stradario.Strutture
     public class Nodo : Sicurezza
     {
         [Key]
-        public int idNodo { get; set; }
+        public uint idNodo { get; set; }
         [Required]
         public string nome { get; set; }
-        public double X { get; set; }
-        public double Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public override string ToString()
         {
             return this.nome;
         }
 
-        public double CalcolaDistanza(Nodo target)
+        public float CalcolaDistanza(Nodo target)
         {
             // calcolare la distanza tra i due nodi
-            double deltaX = this.X - target.X;
-            double deltaY = this.Y - target.Y;
-            double distanza = Math.Sqrt(deltaX * deltaX + deltaY * deltaY);
+            float deltaX = this.X - target.X;
+            float deltaY = this.Y - target.Y;
+            float distanza = (float)Math.Sqrt(deltaX * deltaX + deltaY * deltaY);
             return distanza;
         }
     }
